// src/app/pages/cadastrar-funcionario/model/funcionario-model.ts
export interface FuncionarioModel{
    id?: number;
    nome?: string;
    cpf?: string;
    coren?: string;
    endereco?: EnderecoModel;
    email?: string;
    telefone?: string;
    telefoneEmergencia?: string;
    senha?: string;
}

export interface EnderecoModel {
    logradouro?: string;
    numero?: string;
    complemento?: string;
    bairro?: string;
    cidade?: string;
    estado?: string;
    cep?: string;
}